refactor(calcDate): Seperate calcData function
